added rpyc config
